skills/stealth-scraper-core/requirements.txt
fastapi
uvicorn
playwright
playwright-stealth
python-dotenv
requests
